+++ pmd-ant/src/main/java/net/sourceforge/pmd/ant/internal/PMDTaskImpl.java
@@ -5,6 +5,7 @@
 package net.sourceforge.pmd.ant.internal;
 
 import java.nio.charset.Charset;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -16,6 +17,7 @@ import org.apache.tools.ant.DirectoryScanner;
 import org.apache.tools.ant.Project;
 import org.apache.tools.ant.types.FileSet;
 import org.apache.tools.ant.types.Path;
+import org.apache.tools.ant.types.Resource;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -26,6 +28,7 @@ import net.sourceforge.pmd.PmdAnalysis;
 import net.sourceforge.pmd.RulePriority;
 import net.sourceforge.pmd.RuleSetLoader;
 import net.sourceforge.pmd.ant.Formatter;
+import net.sourceforge.pmd.ant.InternalApiBridge;
 import net.sourceforge.pmd.ant.PMDTask;
 import net.sourceforge.pmd.ant.SourceLanguage;
 import net.sourceforge.pmd.internal.Slf4jSimpleConfiguration;
@@ -54,7 +57,7 @@ public class PMDTaskImpl {
     private Project project;
 
     public PMDTaskImpl(PMDTask task) {
-        configuration.addRelativizeRoots(task.getRelativizeRoots());
+        configuration.addRelativizeRoots(getRelativizeRoots(task));
         if (task.getSuppressMarker() != null) {
             configuration.setSuppressMarker(task.getSuppressMarker());
         }
@@ -96,6 +99,17 @@ public class PMDTaskImpl {
 
         project = task.getProject();
     }
+
+    private static List<java.nio.file.Path> getRelativizeRoots(PMDTask pmdTask) {
+        List<java.nio.file.Path> paths = new ArrayList<>();
+        for (Path path : pmdTask.getRelativizePathsWith()) {
+            for (Resource resource : path) {
+                paths.add(Paths.get(resource.toString()));
+            }
+        }
+        return paths;
+    }
+
 
     private void doTask() {
         setupClassLoader();
@@ -161,7 +175,7 @@ public class PMDTaskImpl {
             renderers.add(reportSizeListener);
             for (Formatter formatter : formatters) {
                 project.log("Sending a report to " + formatter, Project.MSG_VERBOSE);
-                renderers.add(formatter.newListener(project));
+                renderers.add(InternalApiBridge.newListener(formatter, project));
             }
             return GlobalAnalysisListener.tee(renderers);
         } catch (Exception e) {